Fix #2, better Chinese word predication
Use char-offset-in-word instead of char-at-point for precise match,
below is a short example for this fix:

Assuming some Chinese text in a buffer and point at the 2nd 中
-- buffer --
中国人使用中文
------------

the result of Chinese word segmentation:
=> 中国 人 使用 中文
   1 2  3 4 5  6 7

hence:
char-at-point = 中
char-offset-in-word = 6 (by using this, we can distinguish 中文 from 中
国)

In this commit jieba is *required* for Chinese word segmentation. In the
future, we should provide a custom option:
1. scws
2. jieba
3. none
so that it (使用中文分词) becomes a optional feature.